% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RESULTS}
\alias{RESULTS}
\title{RESULTS}
\format{
A data frame with 1320 rows and 5 variables:
\describe{
\item{\code{r_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{title}}{character COLUMN_DESCRIPTION}
\item{\code{citation}}{character COLUMN_DESCRIPTION}
\item{\code{snippet}}{character COLUMN_DESCRIPTION}
}
}
\usage{
RESULTS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
