% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RESULTS_LOG}
\alias{RESULTS_LOG}
\title{RESULTS_LOG}
\format{
A data frame with 106 rows and 6 variables:
\describe{
\item{\code{rl_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{search_term}}{character COLUMN_DESCRIPTION}
\item{\code{processed_search_term}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{max_return_size}}{double COLUMN_DESCRIPTION}
\item{\code{results_count}}{double COLUMN_DESCRIPTION}
}
}
\usage{
RESULTS_LOG
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
